refactor: adjust helpers & exports
