+++ packages/safe-ds-lang/src/helpers/collections.ts
