+++ packages/safe-ds-lang/src/helpers/locations.ts
