+++ packages/safe-ds-lang/src/helpers/strings.ts
