+++ packages/safe-ds-lang/src/language/generation/safe-ds-python-generator.ts
@@ -19,7 +19,7 @@ import {
 import path from 'path';
 import { SourceMapGenerator, StartOfSourceMap } from 'source-map';
 import { TextDocument } from 'vscode-languageserver-textdocument';
-import { groupBy } from '../../helpers/collectionUtils.js';
+import { groupBy } from '../../helpers/collections.js';
 import { SafeDsAnnotations } from '../builtins/safe-ds-annotations.js';
 import {
     isSdsAbstractResult,
