+++ packages/safe-ds-lang/src/language/index.ts
@@ -1,7 +1,18 @@
+// Services
 export type { SafeDsServices } from './safe-ds-module.js';
-export * as ast from './generated/ast.js';
-export { SafeDsLanguageMetaData } from './generated/module.js';
-export { startLanguageServer } from './main.js';
 export { createSafeDsServices, createSafeDsServicesWithBuiltins } from './safe-ds-module.js';
-export { SAFE_DS_FILE_EXTENSIONS } from './helpers/fileExtensions.js';
-export { getModuleMembers } from './helpers/nodeProperties.js';
+
+// Language Server
+export { startLanguageServer } from './main.js';
+
+// Language Metadata
+export { SafeDsLanguageMetaData } from './generated/module.js';
+
+// AST
+import * as generatedAst from './generated/ast.js';
+import * as nodeProperties from './helpers/nodeProperties.js';
+
+export const ast = { ...generatedAst, ...nodeProperties };
+
+// Location
+export { locationToString, positionToString, rangeToString } from '../helpers/locations.js';
