+++ packages/safe-ds-lang/src/language/partialEvaluation/safe-ds-partial-evaluator.ts
@@ -1,5 +1,5 @@
 import { AstNode, AstNodeLocator, getContainerOfType, getDocument, WorkspaceCache } from 'langium';
-import { isEmpty } from '../../helpers/collectionUtils.js';
+import { isEmpty } from '../../helpers/collections.js';
 import {
     isSdsArgument,
     isSdsAssignee,
