+++ packages/safe-ds-lang/src/language/typing/safe-ds-type-computer.ts
@@ -1,5 +1,5 @@
 import { AstNode, AstNodeLocator, getContainerOfType, getDocument, stream, WorkspaceCache } from 'langium';
-import { isEmpty } from '../../helpers/collectionUtils.js';
+import { isEmpty } from '../../helpers/collections.js';
 import {
     isSdsAnnotation,
     isSdsArgument,
