+++ packages/safe-ds-lang/src/language/validation/other/types/literalTypes.ts
@@ -3,7 +3,7 @@ import { ValidationAcceptor } from 'langium';
 import { getLiterals } from '../../../helpers/nodeProperties.js';
 import { SafeDsServices } from '../../../safe-ds-module.js';
 import { EvaluatedNode } from '../../../partialEvaluation/model.js';
-import { isEmpty } from '../../../../helpers/collectionUtils.js';
+import { isEmpty } from '../../../../helpers/collections.js';
 
 export const CODE_LITERAL_TYPE_DUPLICATE_LITERAL = 'literal-type/duplicate-literal';
 export const CODE_LITERAL_TYPE_LIST_LITERAL = 'literal-type/list-literal';
