+++ packages/safe-ds-lang/src/language/validation/other/types/namedTypes.ts
@@ -2,8 +2,8 @@ import { SdsNamedType } from '../../../generated/ast.js';
 import { ValidationAcceptor } from 'langium';
 import { SafeDsServices } from '../../../safe-ds-module.js';
 import { getTypeArguments, getTypeParameters } from '../../../helpers/nodeProperties.js';
-import { duplicatesBy } from '../../../../helpers/collectionUtils.js';
-import { pluralize } from '../../../../helpers/stringUtils.js';
+import { duplicatesBy } from '../../../../helpers/collections.js';
+import { pluralize } from '../../../../helpers/strings.js';
 
 export const CODE_NAMED_TYPE_DUPLICATE_TYPE_PARAMETER = 'named-type/duplicate-type-parameter';
 export const CODE_NAMED_TYPE_POSITIONAL_AFTER_NAMED = 'named-type/positional-after-named';
