+++ packages/safe-ds-lang/src/language/validation/style.ts
@@ -1,5 +1,5 @@
 import { ValidationAcceptor } from 'langium';
-import { isEmpty } from '../../helpers/collectionUtils.js';
+import { isEmpty } from '../../helpers/collections.js';
 import {
     isSdsEnumVariant,
     isSdsWildcard,
