+++ packages/safe-ds-lang/tests/language/typing/type computer/safe-ds-type-computer.test.ts
@@ -2,7 +2,7 @@ import { describe, it } from 'vitest';
 import { createSafeDsServicesWithBuiltins } from '../../../../src/language/index.js';
 import { NodeFileSystem } from 'langium/node';
 import { AssertionError } from 'assert';
-import { locationToString } from '../../../helpers/location.js';
+import { locationToString } from '../../../../src/helpers/locations.js';
 import { createTypingTests } from './creator.js';
 import { getNodeByLocation } from '../../../helpers/nodeFinder.js';
 import { loadDocuments } from '../../../helpers/testResources.js';
